refactor: share cp tf32 helpers
